fix(hooks): remove the bare-reference link blocking gate
The PreToolUse/Stop bare-reference link gate over-blocked routine commands
([over-block report](https://github.com/souliane/teatree/issues/107)) — most
recently read-only diagnostic commands — by hard-denying any user-facing
publish citing a bare issue/MR reference. The deterministic, code-only
reference linkifier (`core/reference_linkifier.py`) auto-converts bare refs to
clickable links at the send chokepoint, so the blocking gate is redundant and
its over-block recurred without a clean fix in sight.

Removed:
- `handle_bare_reference_pretool` / `handle_bare_reference_stop` and their
  PreToolUse / Stop registrations in `hook_router.py`.
- `src/teatree/hooks/bare_reference_scanner.py` (the gate module). Its
  detection primitives (the regexes + `find_bare_references`), still needed by
  the review-findings leak neutralizer, move into `core/review_findings.py`,
  its only remaining consumer.
- `src/teatree/hooks/publish_destination_kind.py` — orphaned once the gate
  (its only consumer) is gone — plus its test.
- The gate's read-vs-write regression check + corpus entry, the gate's test
  file, and its liveness / never-lockout corpus entries.

The linkifier is kept intact — it is the replacement the removal relies on.
BLUEPRINT and factory-architecture prose updated to describe linkify-only.
