+++ BLUEPRINT.md
@@ -214,7 +214,7 @@ Request parameters are grouped into frozen `slots=True` dataclasses (`PullReques
 
 **Destination-routed post/react ([#1750](https://github.com/souliane/teatree/issues/1750)).** `t3 notify post`/`react` pick the token by *destination* via `SlackBotBackend._route_token`: user's own DM → bot; colleague/channel → personal `xoxp`.
 
-**Deterministic reference linkifier.** The clickable-references rule (a bare `#N` / `!N` on a user-facing surface must become a clickable link) is enforced *in code* by `core/reference_linkifier.py`, not by asking the model. `ReferenceResolver` resolves DB-first (`PullRequest`, `Ticket.issue_url`) then constructs from the active repo context (overlay `code_host` + git-remote slug); the overlay hooks `resolve_mr_token` / `resolve_issue_token` default to it. `linkify` (idempotent; skips linked refs, inline code, fenced blocks; leaves unresolvable refs alone) runs at the Slack chokepoint (`SlackReplier._deliver`). Forge posts are NOT linkified (forge auto-links bare ids — the gate's `EXTERNAL_FORGE` carve-out). The bare-reference gate (`hooks/bare_reference_scanner.py`) is the fallback for unresolved refs and the agent's own chat output.
+**Deterministic reference linkifier.** The clickable-references rule (a bare `#N` / `!N` on a user-facing surface must become a clickable link) is enforced *in code* by `core/reference_linkifier.py`, not by asking the model. `ReferenceResolver` resolves DB-first (`PullRequest`, `Ticket.issue_url`) then constructs from the active repo context (overlay `code_host` + git-remote slug); the overlay hooks `resolve_mr_token` / `resolve_issue_token` default to it. `linkify` (idempotent; skips linked refs, inline code, fenced blocks; leaves unresolvable refs alone) runs at the Slack chokepoint (`SlackReplier._deliver`). Forge posts are NOT linkified (forge auto-links bare ids). It is the sole mechanism for the rule — the former PreToolUse/Stop bare-reference *blocking* gate was removed (it over-blocked routine commands and asked the model to rewrite refs non-deterministically); the linkifier rewrites in code with no block and no model round-trip.
 
 **Sync ABC (`core/sync.py`).** `SyncBackend` is an ABC with `is_configured(overlay)` and `sync(overlay) → SyncResult`. Implementations: `GitHubSyncBackend`, `GitLabSyncBackend`. Both consume `CodeHostBackend` — platform-specific code lives only in the Protocol implementation, not in sync logic.
 
@@ -276,7 +276,6 @@ Egress-leak gate family — one doctrine, several entry points, each named here 
 
 - **Public-repo diff privacy-scan** (#685, #730): pre-push hook `refuse-public-push-with-leak.sh` runs `t3 tool privacy-scan` over the pushed diff + commit messages when `origin` is PUBLIC, blocking emails / home paths / private IPs / keys / internal hostnames / banned terms. It fails CLOSED on a genuine finding (a dedicated exit code) but fails OPEN on any other non-zero so a scanner crash cannot wedge every push (#126).
 - **Banned-terms posting gate** (#1415, `teatree.hooks.banned_terms_scanner`): the `PreToolUse` non-commit sibling, scanning gh/glab post bodies.
-- **Bare-reference link gate** (#1530, `teatree.hooks.bare_reference_scanner`): **destination-KIND-aware** (`publish_destination_kind`) — HARD-denies a USER-FACING body citing an unlinked bare ref, but EXEMPTS external-forge posts (the forge auto-renders the ref) and verbatim quoted blocks. Fail-safe toward user-facing; a `Stop` sibling soft-warns on final chat text.
 - **Pre-dispatch quote-scanner** (#1401, `handle_dispatch_prompt_quote_scanner`): the dispatch-boundary companion to #1213 — denies only a HIGH user-voice/PII match in an `Agent`/`Task` prompt, with a `[quote-ok: <reason>]` opt-out, so a verbatim quote cannot leak downstream.
 - **Diff comment detectors** (added-lines-only): `code_comment_self_reference` (#1465) flags bookkeeping self-references inside comment syntax; `code_comment_density` (#1538) is the commit-side half of the near-zero-comments rule (dispatch-side half in the sub-agent prompt preambles, #1532). Both **fail-open per detector** (#1536). The density logic is also a standalone reusable gate (#1369, `t3 tool comment-density`) shared by the pre-push hook and the `comment-density-gate` CI job (re-run on the PR-vs-base diff so a prek bypass still trips). A golden corpus (`tests/test_comment_density_gate.py`) and a trajectory eval pin must-DENY symmetric with must-ALLOW as the anti-vacuous proof. The push-stage gates (`code_comment_density`, `doc-update-gate`, `pytest-fast`, `ensure-pr`, the leak gate) are `stages: [push]` — skipped by a bare `prek run --all-files` but re-run by CI — so `t3 tool verify-gates` runs both stages and returns the combined exit code; skills and the builder prompt point at it so local green == CI green.
 - **Full-tree banned-brand backstop** (#1570, `core.banned_terms_tree`, CLI `t3 banned-terms scan-tree`, CI job `banned-terms-tree`): catches what the diff/payload gates structurally cannot — a brand ALREADY committed, invisible to any post-landing diff — by scanning every tracked file's content with an underscore-tolerant boundary.
