+++ docs/blueprint/factory-architecture.md
@@ -204,7 +204,7 @@ Three further mechanizable gates in the same family, each replacing a prose-only
 - **Loop self-pump Stop gate.** `handle_loop_self_pump` (`hook_router.py` `Stop` chain) is the §5.6 invariant 2 enforcement: the singleton loop owner ticks the next `t3 loop tick` on its own session, so the cron-armed cadence never has to race with the interactive session for the lease.
 - **Plan-gate on workspace edits (gate 16, [#1133](https://github.com/souliane/teatree/issues/1133)).** `handle_enforce_plan_gate` (`hook_router.py` `PreToolUse`) denies `Edit`/`Write` on any file under `$T3_WORKSPACE_DIR` when the session has neither invoked `/plan` (recorded in `<session>.plan-invocations` by `handle_track_plan_invocation` on every `Skill` tool call whose `skill` final segment is `teatree-plan`) nor `Read` the touched file (recorded in `<session>.workspace-reads` by `handle_track_workspace_source_read` for workspace files only — out-of-workspace reads are NOT recorded, so an unrelated `~/.zshrc` read cannot authorize a workspace `Edit`). Outside the workspace root (`~/.zshrc`, `~/.claude/`, agent memory) the gate passes through silently. Opt-in per overlay via `[overlays.<name>] plan_gate = true` in `~/.teatree.toml` (default `False`): when ANY overlay has it `true` the gate becomes active for workspace-scoped edits on this machine. This codifies the prose-only "plan before any implementation dispatch" directive at the tool layer; the deny message names the file, the rule, and the missing prerequisite so the agent knows the remediation in-line.
 - **Banned-terms posting gate ([#1415](https://github.com/souliane/teatree/issues/1415)).** `handle_banned_terms_pretool` (`hook_router.py` `PreToolUse`, `teatree.hooks.banned_terms_scanner`) is the sibling of the [#1213](https://github.com/souliane/teatree/issues/1213) quote-scanner gate: the commit-only `check-banned-terms.sh` pre-commit hook runs only on `git commit`, missing every non-commit write to a public surface (`gh issue/pr create|edit|comment`, `glab mr|issue note|create`, the `gh api` / `glab api` REST paths) — exactly where overlay/customer terms have leaked. The gate reuses the #1213 `teatree.hooks._command_parser` publish-surface detection + body extraction (`--body`, `--body-file`, `-d`/`--field` JSON, `-m`, heredocs), then delegates the *matching* to the same `check-banned-terms.sh` against the `~/.teatree.toml` term list — no new term config, no reimplemented matching. A match `deny`s the tool call naming the term; the override is `--allow-banned-term` (first segment), a leading `ALLOW_BANNED_TERM=1` env-assignment token, or the `ALLOW_BANNED_TERM=1` process/tool env. Fails open on a missing config/script or subprocess error. Destination-aware via `internal_publish_namespaces` unioned with `private_repos` ([#1672](https://github.com/souliane/teatree/issues/1672)); both gates scan PUBLIC targets only (`teatree.hooks.publish_destination`), the SKIP decided per top-level segment so a chained / command-substituted public post or a raw-REST `gh api`/`glab api` segment is still SCANNED. A SECRET blocks on every surface before any skip/override; a bare `git commit` whose effective repo (cd/pushd + `.git` walk-up) is inside no git repo fails OPEN (local-only), but a PUBLIC-repo commit still hard-blocks.
-- **Bare-reference link gate ([#1530](https://github.com/souliane/teatree/issues/1530)).** `teatree.hooks.bare_reference_scanner`, destination-KIND-aware (`publish_destination_kind`): a `PreToolUse` HARD-`deny` on a USER-FACING bare `#NNNN`/`!NNNN`/`ts` not in a clickable link (+ a `Stop` SOFT warning). An EXTERNAL-FORGE post and a bare URL are exempt.
+- **Clickable-reference rule.** Enforced *in code* by the deterministic linkifier (`core/reference_linkifier.py`, [#1845](https://github.com/souliane/teatree/issues/1845)), which rewrites a bare `#NNNN`/`!NNNN` on a user-facing surface into a clickable link at the Slack chokepoint. The former PreToolUse/Stop bare-reference *blocking* gate ([#1530](https://github.com/souliane/teatree/issues/1530)) was removed: it over-blocked routine commands ([#107](https://github.com/souliane/teatree/issues/107)) and asked the model to rewrite refs non-deterministically; the linkifier replaces it with no block and no model round-trip.
 - **Skill-loading-on-task-create gate (gate 17, [#1488](https://github.com/souliane/teatree/issues/1488)).** `handle_enforce_skill_loading_on_task_create` rides the `TaskCreated` event — the one seam the harness Workflow/Task fan-out vehicle (`ultracode`, teammate spawns) does NOT bypass (unlike `PreToolUse`, where its sibling `handle_enforce_skill_loading` lives), closing the loophole that let a bespoke review workflow run instead of `/t3:review`. It forces the matching teatree skill + transitive companions (`lifecycle_for_task_text(task_description)`, fail-open on stale names) onto the fanned-out task via the `{"continue": false, …}` deny envelope, enforcing skill-loading ONLY so ultracode keeps maximal fan-out room (the `TEATREE_PLAN_GATE_WINDOW_MINUTES = 0` sentinel relaxes gate 16's window for multi-wave fan-outs). Fail-open / self-rescue (never lock out): `[teatree] skill_loading_gate_enabled` kill-switch, `[skip-skill-gate: <reason>]` per-call token, and `t3 <overlay> gate skill-loading disable` — modelled on gate 2. Detail: handler docstring + `hooks/CLAUDE.md`.
 - **No-commit sub-agent recorder (gate 18, [#1205](https://github.com/souliane/teatree/issues/1205)).** `handle_subagent_stop_no_commit` (`SubagentStop`) records a `terminated_without_commit` signal when `teatree.hooks.no_commit_detector` finds the sub-agent's worktree (harness `cwd`) on a work branch with 0 commits ahead of base. A DETECTOR, not a deny; fail-open. Detail: handler docstring + `hooks/CLAUDE.md`.
