+++ src/teatree/eval/README.md
@@ -201,7 +201,6 @@ class, where it is pinned, and the originating fix:
 |---|---|---|
 | migration-fork / multiple leaf nodes | `regression_corpus` (graph leaf count) | [#1721](https://github.com/souliane/teatree/pull/1721) |
 | branch-currency §940 (conflict-only, never behind-only) | `regression_corpus` | [#1719](https://github.com/souliane/teatree/pull/1719) |
-| bare-reference gate over-block on read-only `gh api` | `regression_corpus` | [#1535](https://github.com/souliane/teatree/pull/1535) |
 | substrate-merge human-authorize floor | `regression_corpus` (merge precondition) | [#1498](https://github.com/souliane/teatree/pull/1498) |
 | substrate-merge full-autonomy carve-out | `regression_corpus` (merge precondition) | [#1748](https://github.com/souliane/teatree/issues/1748) |
 | maker≠checker at merge time | `regression_corpus` (merge precondition) | [#1601](https://github.com/souliane/teatree/pull/1601) |
